<root>/ApplicationAMBCPooling is a sample application to show the best practice of using as few as possible root application modules in an ADF Application. First usecase utilizes 5 root application modules and has a reset user session button(while there is a default configuration of 15 pooled db connections). Clicking the button more than 3 times results in an exception.
Second usecase wrappes the 5 application modules in one root module. This works fine after subsequent user session resets.

Also is followed the lack of parallelism with default settings by adding a timeout of 20sec to the query operation(result is that each of the 5 tables shown in the UI will wait step by step for the model to execute the long time taking queries)