<root>/ApplicationContextualEvents
Use Case:  send a parameter from one page fragment to another.

For this POC we will use 2 distinct task flows called first-region and second-region. Each task flow consists of a single view , firstChild  respectively secondChild. These two task flows are inserted into the index page as regions. The goal is to send a parameter (email) from the first child onto the second.

For this in the first page fragment we will have an ADF form based on the Employees table in the HR schema and every time we modify the value of the Email field we want to capture the new value in an output text on the 2nd fragment.